+++ library/src/commonTest/kotlin/io/github/ezer_mackenzie/proximitytransfer/core/transfer/ProtocolTransferTest.kt
@@ -1,11 +1,13 @@
 package io.github.ezer_mackenzie.proximitytransfer.core.transfer
 
+import io.github.ezer_mackenzie.proximitytransfer.core.integrity.IntegrityVerificationException
 import io.github.ezer_mackenzie.proximitytransfer.core.transport.memory.MemoryTransport
 import kotlinx.coroutines.async
 import kotlinx.coroutines.test.runTest
 import kotlin.random.Random
 import kotlin.test.Test
 import kotlin.test.assertContentEquals
+import kotlin.test.assertFailsWith
 
 class ProtocolTransferTest {
     @Test
@@ -23,5 +25,16 @@ class ProtocolTransferTest {
 
             assertContentEquals(expected, received.await(), "Payload size: $size")
         }
+    }
+
+    @Test
+    fun rejectsPayloadCorruptedInTransit() = runTest {
+        val (senderTransport, receiverTransport) = MemoryTransport.createPair()
+        val sender = ProtocolSender(CorruptingConnection(senderTransport.open()))
+        val receiver = ProtocolReceiver(receiverTransport.open())
+
+        sender.send(byteArrayOf(1, 2, 3))
+
+        assertFailsWith<IntegrityVerificationException> { receiver.receive() }
     }
 }